Initial Framework
This framework is build-able within Linux using the "arm-none-eabi-gcc" compiler.

This framework contains the minimal set of FreeRTOS files needed to run the
features included.  The FreeRTOS files are in the "rtos" directory in the same
directory structure as provided in the FreeRTOS release.

This framework has been tested on Freescale FRDM-KL25Z boards.

The main interface to this framework is via the USB UART via the OpenSDA port.
The interface consists of a serial command line which contains a number of
commands to interact with some of the hardware peripherals available on the
FRDM-KL25Z board.  Type "help" to see the list of commands.