Final USB Cleanup
The change adds the final USB device driver cleanup.

The USB driver is now fairly cleanly layered.  The lowest layer,
usbDev.c/h hides the majority of the interactions with the hardware.
The CDC layer, usbCdc.c/h, encapsulates the CDC functionality for
an emulated serial device.  The app layer, usbTask.c/h, handles the
basic loop back behavior.

The interfaces between the layers are now based on higher layer
initializations of lower layers and a series of callback mechanisms.
Callbacks can be argued one way or the other, but they make for
fairly clean layers.  The downside is that the callbacks are done
within ISR contexts, which can be troublesome if not dealt with
carefully.  In all cases, callbacks should do very little, especially
at the app level.  My plans are for simply buffering of the data,
if data is given, and signalling the task of the event.  Nothing
more.