Added USB Task Rx Buffer
This change adds a circular buffer utility (utilsBuf.c/h) and uses
this to create a USB receive data buffer within the USB task.  The
USB task now pays attentions to the Tx Done event and will only
send more Tx data when the USB driver is ready for more.