Console On KL25Z USB Port
This change adds an option to move the command line console from the
debug port to the KL25Z USB port.  This option is currently enabled
in the baseline, but can be controlled via the USB_CONSOLE_ENABLED
define in the "frdm/frdmCfg.h" file.  When the console on the KL25Z
USB port is not enabled, the console is on the debug USB port and
the KL25Z USB port is in echo mode.