Adding support for controlling servo motors
This change adds a PIT based servo motor driver and console commands
to call the servo driver.

This code compiles but has not been tested yet.

The servo motor driver uses a single PIT timer to provide the timing
and is expected to support up to 16 servo motors simultaneously.
Each servo is given a slot of time in which the produce a timing pulse,
if opened and enabled.  The channels are handled in a round-robin
fashion such that the total time between two consecutive pulses from
the same channel are 40ms apart.